Moving things around. Keyword completion works, but needs better handling
