+++ paket.lock
@@ -1,28 +1,26 @@
 NUGET
   remote: http://nuget.org/api/v2
   specs:
-    DotNetZip (1.9.3)
-    FAKE (3.15.0)
-    FSharp.Compiler.Service (0.0.82)
+    FAKE (3.31.2)
+    FSharp.Compiler.Service (0.0.89)
     FSharp.Core (3.1.2.1)
-    FSharp.Formatting (2.7.4)
-      FSharp.Compiler.Service (>= 0.0.82)
-      FSharpVSPowerTools.Core (>= 1.7.0)
-    FSharpVSPowerTools.Core (1.7.0)
-      FSharp.Compiler.Service (>= 0.0.81)
-    Microsoft.Bcl (1.1.9)
+    FSharp.Formatting (2.9.5)
+      FSharp.Compiler.Service (>= 0.0.87)
+      FSharpVSPowerTools.Core (1.8.0)
+    FSharpVSPowerTools.Core (1.8.0)
+      FSharp.Compiler.Service (>= 0.0.87)
+    Microsoft.Bcl (1.1.10)
       Microsoft.Bcl.Build (>= 1.0.14)
     Microsoft.Bcl.Build (1.0.21)
-    Microsoft.Net.Http (2.2.28)
-      Microsoft.Bcl (>= 1.1.9)
+    Microsoft.Net.Http (2.2.29)
+      Microsoft.Bcl (>= 1.1.10)
       Microsoft.Bcl.Build (>= 1.0.14)
     Newtonsoft.Json (6.0.8)
-    Octokit (0.6.2)
+    Octokit (0.10.0)
       Microsoft.Net.Http
-    Paket.Core (0.27.8)
-      DotNetZip (>= 1.9.3)
-      FSharp.Core (>= 3.1.2.1)
-      Newtonsoft.Json (>= 6.0.8)
+    Paket.Core (1.3.30)
+      FSharp.Core
+      Newtonsoft.Json
     VSSDK.ComponentModelHost (12.0.4)
       VSSDK.IDE.12 (>= 12.0.4 < 13.0.0)
     VSSDK.ComponentModelHost.12 (12.0.4)
@@ -86,5 +84,5 @@ NUGET
 GITHUB
   remote: fsharp/FAKE
   specs:
-    modules/Octokit/Octokit.fsx (0d82fea265404d0339744fbe10617fb2737a0c87)
+    modules/Octokit/Octokit.fsx (90eb5c7821ee95a7902d2d88ca9773561b9113c9)
       Octokit
